make MuxTable
