+++ Cargo.toml
@@ -41,8 +41,8 @@ serde_arrays = "0.2.0"
 sha2 = { version = "0.10.9" }
 
 # Uncomment for debugging with https://github.com/ed255/plonky2/ at branch `feat/debug`.  The repo directory needs to be checked out next to the pod2 repo directory.
-[patch."https://github.com/0xPARC/plonky2"]
-plonky2 = { path = "../plonky2/plonky2" }
+# [patch."https://github.com/0xPARC/plonky2"]
+# plonky2 = { path = "../plonky2/plonky2" }
 
 [dev-dependencies]
 pretty_assertions = "1.4.1"
@@ -62,5 +62,5 @@ examples = []
 disk_cache = ["directories", "minicbor-serde"]
 mem_cache = []
 
-[profile.release]
-debug = true
+# [profile.release]
+# debug = true
