+++ src/backends/plonky2/circuits/mod.rs
@@ -1,5 +1,7 @@
 pub mod common;
+pub mod hash;
 pub mod mainpod;
 pub mod metrics;
+pub mod mux_table;
 pub mod signedpod;
 pub mod utils;
